supervisor: fixed data type of message counters in graph nodes structures

+++ graph.h
@@ -61,7 +61,7 @@ typedef struct graph_node_output_interface_s graph_node_output_interface_t;
 struct graph_node_input_interface_s {
    interface_t *              ifc_struct; ///< Pointer to interface_t structure of running_module_t struct.
    graph_node_t *                parent_node; ///< Pointer to parent node.
-   int                     message_counter; ///< Received messages counter of this node input interface.
+   uint64_t                     message_counter; ///< Received messages counter of this node input interface.
    int                     node_interface_port; ///< Port of this node interface.
    char *                  unix_sock_interface_port;
    graph_node_output_interface_t * node_interface_output_ifc; ///< Pointer to connected output interface of another module.
@@ -71,7 +71,7 @@ struct graph_node_input_interface_s {
 struct graph_node_output_interface_s {
    interface_t            *ifc_struct; ///< Pointer to interface_t structure of running_module_t struct.
    graph_node_t           *parent_node; ///< Pointer to parent node.
-   unsigned int                   message_counter; ///< Sent messages counter of this node output interface.
+   uint64_t                   message_counter; ///< Sent messages counter of this node output interface.
    unsigned int                   node_children_counter; ///< Node children counter.
    graph_node_input_interface_t **node_children; ///< Array of connected node input interfaces of another nodes (node children).
    int                     node_interface_port; ///< Port of this node interface.
